# pkg/policy/api: Optimize FQDNSelector String()